move errors to a separate file
